add unit tests

+++ Microsoft.Azure.Cosmos/src/CosmosElements/Telemetry/ReportPayload.cs
@@ -138,7 +138,7 @@ namespace Microsoft.Azure.Cosmos.CosmosElements.Telemetry
             return isequal;
         }
 
-        internal void SetAggregators(LongConcurrentHistogram histogram, int adjustment = 1)
+        internal void SetAggregators(LongConcurrentHistogram histogram, long adjustment = 1)
         {
             this.MetricInfo.SetAggregators(histogram, adjustment);
         }
